Finished assigning subject to section v2

+++ class/Administration.php
@@ -843,6 +843,85 @@ class Administration extends Dbconfig
         );
     }
 
+//    public function getSectionSubClassInfo()
+//    {
+//        session_start();
+//        $code = $_GET['code'];
+//        $programs = [];
+//        $recommended = [];
+//        $sy_id = $_SESSION['sy_id'];
+//
+//        $grd_level = mysqli_fetch_row($this->query("SELECT grd_level FROM section WHERE section_code = '$code';"))[0];
+//        $result = $this->query("SELECT DISTINCT(prog_code) FROM enrollment WHERE sy_id = '$sy_id' AND section_code = '$code'; ");
+//        while($row = mysqli_fetch_row($result)) {
+//            $program = $row[0];
+//            $programs[] = ["code" => $program, "link" => "<a href='program.php?prog_code=$program' target='_blank' role='button' class='btn btn-sm btn-primary rounded-pill m-1'>$program</a>"];
+//        }
+//
+//        # retrieve current subjects assigned to the section
+//        $current_sub_code = $this->getCurrentSubjectsOfSection($code, $sy_id); # [BSMATH, ]
+//
+//        # retrieve faculty options
+//        $faculties = $this->listFaculty();
+//        function createFacultySelect ($sub_class_code, $teacher_id, $faculties): string
+//        {
+//            $select = "<select name='subjectClass[$sub_class_code]' class='teacher-select select2--small'>"
+//                        ."<option value=''>-- Select Faculty --</option>";
+//            foreach ($faculties as $faculty) {
+//                $teach_id = (int) $faculty->get_teacher_id();
+//                $select .= "<option value='$teach_id' ". ($teach_id == (int) $teacher_id ? "selected" : "") .">{$faculty->get_name()}</option>";
+//            }
+//            $select .= "</select>";
+//            return $select;
+//        }
+//        $none_display = "style='display: none;'";
+//        foreach($programs as $prog) {
+//            $prog_code = $prog['code'];
+//            $result = $this->query("SELECT sub_code, sub_semester, sub_name FROM sharedsubject JOIN subject USING (sub_code) WHERE prog_code = '$prog_code' AND sy_id = '$sy_id' AND for_grd_level = '$grd_level';");
+//            while($row = mysqli_fetch_assoc($result)) {
+//                $sub_code = $row['sub_code'];
+//                $checked_state = (in_array($sub_code, $current_sub_code) ? "checked" : "");
+//
+//                # get current subject teacher
+//                $teach_row = NULL;
+//                $sub_class_code = NULL;
+//                $teacher_id = NULL;
+//
+//                $teach_result = $this->query("SELECT sc.sub_class_code, sc.teacher_id FROM subjectclass sc
+//                                        JOIN sysub s ON sc.sub_sy_id = s.sub_sy_id
+//                                        LEFT JOIN faculty f ON sc.teacher_id = f.teacher_id WHERE sc.section_code='$code' AND sc.sub_code='$sub_code';");
+//
+//                if (mysqli_num_rows($teach_result) != 0) {
+//                    $teach_row = mysqli_fetch_row($teach_result);
+//                    [$sub_class_code, $teacher_id] = [$teach_row[0], $teach_row[1] ?? NULL];
+//                }
+//
+//
+//                $add_btn_display =  is_null($teacher_id) ? "" : $none_display;
+//                $unassign_display = is_null($teacher_id) ? $none_display : "";
+////                $add_btn_display =  is_null($teacher_id) ? "" : "d-none";
+////                $unassign_display = is_null($teacher_id) ? "d-none" : "";
+//                $action = "<div class='d-flex align-content-center'>"
+//                            . "<button class='btn-success btn btn-sm me-1 action' data-type='add' type='button' $add_btn_display><i class='bi bi-plus-lg'></i></button>"
+//                            . "<button class='btn-danger btn btn-sm me-1 action' data-type='unassign'  data-sub-class-code='$sub_class_code' title='Unassign Faculty' type='button' $unassign_display><i class='bi bi-person-dash-fill'></i></button>"
+//                            . "<span class='w-100'>". createFacultySelect($sub_class_code, $teacher_id, $faculties)."</span>"
+//                        . "</div>";
+//
+//                $recommended[$prog_code][$row['sub_semester']][] = utf8_encode("<tr>"
+//                    ."<td class='text-center'><input class='form-check-input' type='checkbox' name='subjects[$prog_code][]' value='$sub_code' id='$sub_code' $checked_state></td>"
+//                    ."<td><label class='form-check-label' for='$sub_code'>{$row['sub_name']}</label></td>"
+//                    ."<td>$action</td>"
+//                ."</tr>");
+//            }
+//        }
+//
+//        echo json_encode([
+//            "programs"      => $programs,
+//            "recommended"   => $recommended
+//        ]);
+//
+//    }
+
     public function getSectionSubClassInfo()
     {
         session_start();
@@ -862,63 +941,69 @@ class Administration extends Dbconfig
         $current_sub_code = $this->getCurrentSubjectsOfSection($code, $sy_id);
 
         # retrieve faculty options
-        $faculties = $this->listFaculty();
-        function createFacultySelect ($sub_class_code, $teacher_id, $faculties) {
-            $select = "<select name='subjectClass[$sub_class_code]' class='teacher-select select2--small'>
-                        <option value=''>-- Select Faculty --</option>'";
-            foreach ($faculties as $faculty) {
-                $teach_id = $faculty->get_teacher_id();
-                $select .= "<option value='{$teach_id}' ". ($teach_id == $teacher_id ? "selected" : "") .">{$faculty->get_name()}</option>";
-            }
-            $select .= "</select>";
-            return $select;
-        }
+        $none_display = "style='display: none;'";
+        $sub_class_faculty = [];
+        $current_subjects = [];
         foreach($programs as $prog) {
             $prog_code = $prog['code'];
             $result = $this->query("SELECT sub_code, sub_semester, sub_name FROM sharedsubject JOIN subject USING (sub_code) WHERE prog_code = '$prog_code' AND sy_id = '$sy_id' AND for_grd_level = '$grd_level';");
             while($row = mysqli_fetch_assoc($result)) {
                 $sub_code = $row['sub_code'];
-                $checked_state = (in_array($sub_code, $current_sub_code) ? "checked" : "");
+                $sub_semester = $row['sub_semester'];
 
-                # get current subject teacher
-                $teach_row = NULL;
-                $sub_class_code = NULL;
-                $teacher_id = NULL;
-                $teacher_name = NULL;
 
-                $teach_result = $this->query("SELECT sc.sub_class_code, sc.teacher_id FROM subjectclass sc
-                                        JOIN sysub s ON sc.sub_sy_id = s.sub_sy_id
-                                        LEFT JOIN faculty f ON sc.teacher_id = f.teacher_id WHERE sc.section_code='$code' AND sc.sub_code='$sub_code';");
+                if (!in_array($sub_code, $current_subjects)) {
+                    $current_subjects[] = $sub_code;
+//                    print_r($current_subjects);
+//                    echo "<br>";
+//                    echo "$sub_code\n";
+                    $checked_state = (in_array($sub_code, $current_sub_code) ? "checked" : "");
 
-                if (mysqli_num_rows($teach_result) != 0) {
-                    $teach_row = mysqli_fetch_row($teach_result);
-                    [$sub_class_code, $teacher_id] = [$teach_row[0], $teach_row[1] ?? NULL];
+                    # get current subject teacher
+                    $teach_row = NULL;
+                    $sub_class_code = NULL;
+                    $teacher_id = NULL;
+
+                    $teach_result = $this->query("SELECT sc.sub_class_code, sc.teacher_id FROM subjectclass sc
+                                            JOIN sysub s ON sc.sub_sy_id = s.sub_sy_id
+                                            LEFT JOIN faculty f ON sc.teacher_id = f.teacher_id WHERE sc.section_code='$code' AND sc.sub_code='$sub_code';");
+
+                    if (mysqli_num_rows($teach_result) != 0) {
+                        $teach_row = mysqli_fetch_row($teach_result);
+                        $sub_class_code = $teach_row[0];
+                        $teacher_id = $teach_row[1] ?? NULL;
+                        $sub_class_faculty[] = [
+                            'sect_code' => $code,
+                            'sub_code'  => $sub_code,
+                            'sub_teacher' => $teacher_id
+                        ];
+                    }
+
+
+                    $add_btn_display = is_null($teacher_id) ? "" : $none_display;
+                    $unassign_display = is_null($teacher_id) ? $none_display : "";
+                    $action = "<div class='d-flex align-content-center'>"
+                        . "<button class='btn-success btn btn-sm me-1 action' data-type='add' type='button' $add_btn_display><i class='bi bi-plus-lg'></i></button>"
+                        . "<button class='btn-danger btn btn-sm me-1 action' data-type='unassign' data-section='$code' data-sub-code='$sub_code' title='Unassign Faculty' type='button' $unassign_display><i class='bi bi-person-dash-fill'></i></button>"
+                        . "<span class='w-100'><select name='subjectClass[$code][$sub_code]' class='teacher-select select2--small'>"
+                        . "<option value=''>-- Select Faculty --</option></select></span>"
+                        . "</div>";
+
+                    $recommended[$prog_code][$sub_semester][] = [
+                        "checkbox" => "<input class='form-check-input' type='checkbox' name='subjects[$prog_code][]' value='$sub_code' id='$sub_code' $checked_state>",
+                        "subName" => "<label class='form-check-label' for='$sub_code'>{$row['sub_name']}</label>",
+                        "action" => $action
+                    ];
                 }
-
-                $add_btn_display =  is_null($teacher_id) ? "" : "d-none";
-                $unassign_display = is_null($teacher_id) ? "d-none" : "";
-                $action = "<div class='d-flex'>"
-                            . "<button class='btn-success btn btn-sm me-1 action $add_btn_display' data-type='add' type='button'><i class='bi bi-plus-lg'></i></button>"
-                            . "<button class='btn-danger btn btn-sm me-1 action $unassign_display' data-type='unassign'  data-sub-class-code='$sub_class_code' title='Unassign Faculty' type='button'><i class='bi bi-person-dash-fill'></i></button>"
-                            . "<span>". createFacultySelect($sub_class_code, $teacher_id, $faculties)."</span>"
-                            . "</p></div>";
-
-                $recommended[$prog_code][$row['sub_semester']][] = "<tr>
-                    <td class='text-center'><input class='form-check-input' type='checkbox' name='subjects[$prog_code][]' value='$sub_code' id='$sub_code' $checked_state></td>
-                    <td>
-                          <label class='form-check-label' for='$sub_code'>
-                            {$row['sub_name']}
-                          </label>
-                    </td>
-                    <td>$action</td>
-                </tr>";
             }
         }
 
         echo json_encode([
             "programs"      => $programs,
-            "recommended"   => $recommended
+            "recommended"   => $recommended,
+            "currentSubTeacher" => $sub_class_faculty,
         ]);
+
     }
 
     public function listSectionStudentJSON()
@@ -1629,17 +1714,17 @@ class Administration extends Dbconfig
 
     public function listFaculty()
     {
-//        $result = $this->query("SELECT * FROM faculty WHERE teacher_user_no IN (SELECT id_no from user WHERE is_active=1);");
-        $result = $this->query("SELECT * FROM faculty f JOIN user u ON f.teacher_user_no = u.id_no;");
+//        $result = $this->query("SELECT * FROM faculty WHERE teacher_user_no = ANY(SELECT id_no from user WHERE is_active=1 AND user_type = 'FA');");
+        $result = $this->query("SELECT * FROM faculty f JOIN user u ON f.teacher_user_no = u.id_no  WHERE is_active=1;");
         $facultyList = array();
 
         while ($row = mysqli_fetch_assoc($result)) {
             $facultyList[] = new Faculty(
                 $row['teacher_id'],
                 $row['last_name'],
-                $row['middle_name'],
+                $row['middle_name'] ?: "",
                 $row['first_name'],
-                $row['ext_name'],
+                $row['ext_name'] ?: "",
                 $row['birthdate'],
                 $row['age'],
                 $row['sex'],
@@ -2101,7 +2186,8 @@ class Administration extends Dbconfig
                 NULL,
                 NULL,
                 NULL,
-                $row['is_active']
+                $row['is_active'],
+                $row['prog_code']
             );
         }
         return $studentList;
@@ -2334,28 +2420,40 @@ class Administration extends Dbconfig
         $sy_id = $_SESSION['sy_id'];
         # retrieve current subject codes offered in the section
         $current_sub_codes = $this->getCurrentSubjectsOfSection($code, $sy_id);
-        foreach($_POST['subjects'] as $prog_code => $sub_codes) {
-            # delete
-            $to_delete = array_diff($current_sub_codes, $sub_codes);
-            foreach($to_delete as $sub_code_delete) {
-                $this->query("DELETE FROM subjectclass WHERE section_code = '$code' AND sub_code = '$sub_code_delete' AND
-                           sub_sy_id = ANY(SELECT sub_sy_id FROM sysub WHERE sy_id = '$sy_id' AND sub_code = '$sub_code_delete');");
+        if (isset($_POST['subjects'])) {
+            foreach($_POST['subjects'] as $prog_code => $sub_codes) {
+                # delete
+                $to_delete = array_diff($current_sub_codes, $sub_codes);
+                foreach($to_delete as $sub_code_delete) {
+                    $this->query("DELETE FROM subjectclass WHERE section_code = '$code' AND sub_code = '$sub_code_delete' AND
+                               sub_sy_id = ANY(SELECT sub_sy_id FROM sysub WHERE sy_id = '$sy_id' AND sub_code = '$sub_code_delete');");
+                }
+                # add
+                $to_add = array_diff($sub_codes, $current_sub_codes);
+                foreach($to_add as $sub_code_add) {
+                    $sub_sy_id = mysqli_fetch_row($this->query("SELECT sub_sy_id FROM sysub WHERE sy_id = '$sy_id' AND sub_code = '$sub_code_add';"))[0];
+                    $this->query("INSERT INTO subjectclass (sub_sy_id, section_code, sub_code) VALUES ('$sub_sy_id', '$code', '$sub_code_add');");
+                }
             }
-            # add
-            $to_add = array_diff($sub_codes, $current_sub_codes);
-            foreach($to_add as $sub_code_add) {
-                $sub_sy_id = mysqli_fetch_row($this->query("SELECT sub_sy_id FROM sysub WHERE sy_id = '$sy_id' AND sub_code = '$sub_code_add';"))[0];
-                $this->query("INSERT INTO subjectclass (sub_sy_id, section_code, sub_code) VALUES ('$sub_sy_id', '$code', '$sub_code_add');");
-            }
+        } else {
+            # delete all subject class
         }
 
         # update subject faculty
-        foreach($_POST['subjectClass'] as $sub_class_code => $teacher_id) {
-            $teacher_id = empty($teacher_id) ? "NULL" : "'$teacher_id'";
-            echo $teacher_id;
-            echo "<br>";
-//            echo "UPDATE subjectclass SET teacher_id = $teacher_id WHERE sub_class_code = '$sub_class_code';";
-            $this->query("UPDATE subjectclass SET teacher_id = $teacher_id WHERE sub_class_code = '$sub_class_code';");
+        foreach($_POST['subjectClass'] as $section_code => $subjectData) {
+            foreach($subjectData as $subject_code => $teacher_id) {
+                $teacher_id = empty($teacher_id) ? "NULL" : "'$teacher_id'";
+
+                $result = $this->query("SELECT sub_class_code FROM subjectclass WHERE section_code = '$section_code' AND sub_code = '$subject_code';");
+                $query = '';
+                if (mysqli_num_rows($result) != 0) {
+                    $sub_class_code = mysqli_fetch_row($result)[0];
+                    $query = "UPDATE subjectclass SET teacher_id = $teacher_id WHERE sub_class_code = '$sub_class_code'";
+                } else {
+                    $query = "INSERT INTO subjectclass (teacher_id, section_code, sub_code) VALUES ('$teacher_id', '$section_code', '$subject_code');";
+                }
+                $this->query($query);
+            }
         }
     }
